src/main/webapp/app/entities/seller-statistics/route/seller-statistics-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ISellerStatistics } from '../seller-statistics.model';
import { SellerStatisticsService } from '../service/seller-statistics.service';

import { SellerStatisticsRoutingResolveService } from './seller-statistics-routing-resolve.service';

describe('SellerStatistics routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: SellerStatisticsRoutingResolveService;
  let service: SellerStatisticsService;
  let resultSellerStatistics: ISellerStatistics | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(SellerStatisticsRoutingResolveService);
    service = TestBed.inject(SellerStatisticsService);
    resultSellerStatistics = undefined;
  });

  describe('resolve', () => {
    it('should return ISellerStatistics returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSellerStatistics = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSellerStatistics).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSellerStatistics = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultSellerStatistics).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<ISellerStatistics>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSellerStatistics = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSellerStatistics).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
